backend/requirements.txt
cryptography==42.0.7
Django==3.2
django-filter==23.1
django-templated-mail==1.1.1
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
py==1.11.0
pycparser==2.22
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
django-rest-authtoken==2.1.4
Pillow==10.4.0
psycopg2==2.9.9